+++ packages/components/src/index.ts
@@ -25,7 +25,7 @@ export { RouterProvider, type RouterProviderProps } from './components/RouterPro
 export { ScrollArea } from './components/ScrollArea/ScrollArea';
 export { Section } from './components/Section/Section';
 export { SegmentedControl } from './components/SegmentedControl/SegmentedControl';
-export { Select } from './components/Select/Select';
+export { Select } from './components/Select';
 export { Slider } from './components/Slider/Slider';
 export { SplitButton } from './components/SplitButton/SplitButton';
 export { Switch } from './components/Switch/Switch';